scraper-web/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
playwright>=1.40.0
requests>=2.31.0
python-multipart>=0.0.6
